feat(retl): implement Customer.io Audience connection with typed resource support
🔒 Scanned for secrets using gitleaks 8.28.0
